Updated to fit new setup

+++ install-duserver.sh
@@ -12,23 +12,27 @@ function print_success {
 	fi
 }
 
-echo -n "Installing nginx..."
-apt-get -y install nginx > /dev/null
-print_success $?
+if [ ! -e "/etc/init.d/nginx" ]; then
+	echo -n "Installing nginx..."
+	apt-get -y install nginx > /dev/null
+	print_success $?
+fi
 
-echo -n "Adapting configuration..."
-cat <<EOF > /etc/nginx/sites-available/duserver
+echo -n "Adding configuration..."
+cat <<EOF > /etc/nginx/sites-available/duserver.conf
 server {
     root /var/www/modules;
-    index versions.json;
-
-    server_name duserver.*;
+    server_name packages.*;
 
     location / {
         allow all;
     }
 }
 EOF
+print_success $?
+
+echo -n "Enabling configuration..."
+ln -s /etc/nginx/sites-available/duserver.conf /etc/nginx/sites-enabled/ > /dev/null
 print_success $?
 
 if [ ! -d "/var/www" ]; then
@@ -38,34 +42,17 @@ if [ ! -d "/var/www" ]; then
 	chown www-data:www-data /var/www
 fi
 
-if [ ! -d "/var/www/versions" ]; then
-	echo -n "Create root directory for versions..."
-	mkdir /var/www/versions
+if [ ! -d "/var/www/packages" ]; then
+	echo -n "Create root directory for packages..."
+	mkdir /var/www/packages
 	print_success $?
-	chown www-data:www-data /var/www/versions
-	chmod ug+rw /var/www/versions
-fi
-
-if [ ! -d "/var/www/versions/openmrs" ]; then
-	echo -n "Create directory for openmrs versions..."
-	mkdir /var/www/versions/openmrs
-	print_success $?
-	chown www-data:www-data /var/www/versions/openmrs
-	chmod ug+rw /var/www/versions/openmrs
-fi
-
-if [ ! -d "/var/www/versions/androidclient" ]; then
-	echo -n "Create directory for androidclient versions..."
-	mkdir /var/www/versions/androidclient
-	print_success $?
-	chown www-data:www-data /var/www/versions/androidclient
-	chmod ug+rw /var/www/versions/androidclient
+	chown www-data:www-data /var/www/packages
+	chmod ug+rw /var/www/packages
 fi
 
 echo -n "Starting nginx..."
 /etc/init.d/nginx start > /dev/null
 print_success $?
-
 
 #TODO: wget a script that needs to be run when a usb drive is entered
 #TODO: add udev rule file that triggers the script to be run on 'add'
